changed testcases.
