+++ src/functionaltests/java/com/ericsson/ei/rabbitmq/configuration/RabbitMQConfigurationTestSteps.java
@@ -40,7 +40,6 @@ import cucumber.api.java.en.When;
 @Ignore
 @TestPropertySource(properties = {
         "spring.data.mongodb.database: RabbitMQConfigurationTestSteps",
-        "bindingkeys.collection.name: RabbitMQConfigurationTestSteps-bindingKeys",
         "missedNotificationDataBaseName: RabbitMQConfigurationTestSteps-missedNotifications",
         "rabbitmq.exchange.name: RabbitMQConfigurationTestSteps-exchange",
         "rabbitmq.consumerName: RabbitMQConfigurationTestStepsConsumer" })
@@ -55,9 +54,6 @@ public class RabbitMQConfigurationTestSteps extends FunctionalTestBase {
     private static final String ROUTING_KEY_1 = "routing-key-1";
     private static final String ROUTING_KEY_2 = "routing-key-2";
 
-    private static final String BINDING_KEY_1 = "binding-key-1";
-    private static final String BINDING_KEY_2 = "binding-key-2";
-
     private AMQPBrokerManager amqpBroker;
 
     @Autowired
@@ -66,15 +62,6 @@ public class RabbitMQConfigurationTestSteps extends FunctionalTestBase {
 
     @Autowired
     EventHandler eventHandler;
-
-    @Value("${spring.data.mongodb.database}")
-    private String dataBaseName;
-
-    @Value("${bindingkeys.collection.name}")
-    private String collectionName;
-
-    @Autowired
-    private MongoDBHandler mongoDBHandler;
 
     @Given("^We are connected to message bus$")
     public void connect_to_message_bus() throws Exception {
@@ -123,19 +110,6 @@ public class RabbitMQConfigurationTestSteps extends FunctionalTestBase {
                 + missingArguments.toString(), 0, missingArguments.size());
     }
 
-    @When("^add the binding documents to mongoDB$")
-    public void add_the_binding_documents_to_mongoDB() {
-        BasicDBObject dbBinding = insertBinding();
-        assertNotNull(dbBinding);
-    }
-
-    @Then("^compare the binding keys and remove the old binding keys from rabbitMQ and mongoDB$")
-    public void compare_the_binding_keys_and_remove_the_old_binding_keys_from_rabbitMQ_and_mongoDB() {
-        LOGGER.debug("comparing the binding keys to remove the old binding key");
-        ArrayList<String> removedBinding = compareAndRemoveBindings();
-        assertEquals(1, removedBinding.size());
-    }
-
     /**
      * This method collects all the event names of events we will send to the
      * message bus.
@@ -166,35 +140,6 @@ public class RabbitMQConfigurationTestSteps extends FunctionalTestBase {
         rabbitTemplate.setQueue(queueName);
         rabbitTemplate.setRoutingKey(ROUTING_KEY_1);
         return admin;
-    }
-
-    private BasicDBObject insertBinding() {
-        BasicDBObject docInput = new BasicDBObject();
-        docInput.put("destination", "ei-domain.eiffel-intelligence.messageConsumer.durable");
-        docInput.put("destinationType", "QUEUE");
-        docInput.put("exchange", "ei-binding-keys");
-        docInput.put("bindingKeys", BINDING_KEY_1);
-        docInput.put("arg", null);
-        mongoDBHandler.insertDocument(dataBaseName, collectionName, docInput.toString());
-        return docInput;
-    }
-
-    private ArrayList<String> compareAndRemoveBindings() {
-        ArrayList<Binding> listBinding = new ArrayList<Binding>();
-        listBinding.add(new Binding("ei-domain.eiffel-intelligence.messageConsumer.durable", DestinationType.QUEUE,
-                "ei-binding-keys", BINDING_KEY_2, null));
-        List<String> allObjects = mongoDBHandler.getAllDocuments(dataBaseName, collectionName);
-        ArrayList<String> removedBinding = new ArrayList<String>();
-
-        JSONObject dbBinding = new JSONObject(allObjects.get(0));
-        String mongoDbBinding = dbBinding.getString("bindingKeys");
-        if (!(listBinding.contains(mongoDbBinding))) {
-            removedBinding.add(mongoDbBinding);
-            listBinding.remove(0);
-            String condition = "{\"bindingKeys\": /.*" + mongoDbBinding + "/}";
-            mongoDBHandler.dropDocument(dataBaseName, collectionName, condition);
-        }
-        return removedBinding;
     }
 
 }
