+++ src/functionaltests/java/com/ericsson/ei/rabbitmq/connection/RabbitMQTestConnectionSteps.java
@@ -1,18 +1,22 @@
 package com.ericsson.ei.rabbitmq.connection;
 
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
 
 import java.io.File;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.TimeUnit;
 
+import org.json.JSONObject;
 import org.junit.Ignore;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.amqp.core.Binding;
 import org.springframework.amqp.core.BindingBuilder;
 import org.springframework.amqp.core.Queue;
 import org.springframework.amqp.core.TopicExchange;
+import org.springframework.amqp.core.Binding.DestinationType;
 import org.springframework.amqp.rabbit.connection.CachingConnectionFactory;
 import org.springframework.amqp.rabbit.core.RabbitAdmin;
 import org.springframework.amqp.rabbit.core.RabbitTemplate;
@@ -24,9 +28,11 @@ import org.springframework.test.context.TestPropertySource;
 import org.springframework.util.SocketUtils;
 
 import com.ericsson.ei.handlers.EventHandler;
+import com.ericsson.ei.handlers.MongoDBHandler;
 import com.ericsson.ei.handlers.RmqHandler;
 import com.ericsson.ei.utils.AMQPBrokerManager;
 import com.ericsson.ei.utils.FunctionalTestBase;
+import com.mongodb.BasicDBObject;
 
 import cucumber.api.java.en.Given;
 import cucumber.api.java.en.Then;
@@ -35,6 +41,7 @@ import cucumber.api.java.en.When;
 @Ignore
 @TestPropertySource(properties = {
         "spring.data.mongodb.database: RabbitMQTestConnectionSteps",
+        "bindingkeys.collection.name: RabbitMQConfigurationTestSteps-bindingKeys",
         "missedNotificationDataBaseName: RabbitMQTestConnectionSteps-missedNotifications",
         "rabbitmq.exchange.name: RabbitMQTestConnectionSteps-exchange",
         "rabbitmq.consumerName: RabbitMQTestConnectionStepsConsumer" })
@@ -47,6 +54,9 @@ public class RabbitMQTestConnectionSteps extends FunctionalTestBase {
     private static final String EIFFEL_EVENTS = "src/functionaltests/resources/eiffel_events_for_test.json";
 
     private static final String DEFAULT_ROUTING_KEY = "#";
+    
+    private static final String BINDING_KEY_1 = "binding-key-1";
+    private static final String BINDING_KEY_2 = "binding-key-2";
 
     private AMQPBrokerManager amqpBroker;
 
@@ -56,6 +66,15 @@ public class RabbitMQTestConnectionSteps extends FunctionalTestBase {
 
     @Autowired
     EventHandler eventHandler;
+
+    @Value("${spring.data.mongodb.database}")
+    private String dataBaseName;
+
+    @Value("${bindingkeys.collection.name}")
+    private String collectionName;
+
+    @Autowired
+    private MongoDBHandler mongoDBHandler;
 
     @Given("^We are connected to message bus$")
     public void connect_to_message_bus() throws Exception {
@@ -107,6 +126,19 @@ public class RabbitMQTestConnectionSteps extends FunctionalTestBase {
         }
         assertEquals(1, waitListSize);
     }
+    
+    @When("^add the binding documents to mongoDB$")
+    public void add_the_binding_documents_to_mongoDB() {
+        BasicDBObject dbBinding = insertBinding();
+        assertNotNull(dbBinding);
+    }
+
+    @Then("^compare the binding keys and remove the old binding keys from rabbitMQ and mongoDB$")
+    public void compare_the_binding_keys_and_remove_the_old_binding_keys_from_rabbitMQ_and_mongoDB() {
+        LOGGER.debug("comparing the binding keys to remove the old binding key");
+        ArrayList<String> removedBinding = compareAndRemoveBindings();
+        assertEquals(1, removedBinding.size());
+    }
 
     /**
      * This method collects all the event names of events we will send to the
@@ -136,6 +168,35 @@ public class RabbitMQTestConnectionSteps extends FunctionalTestBase {
         rabbitTemplate.setQueue(queueName);
         rabbitTemplate.setRoutingKey(DEFAULT_ROUTING_KEY);
         return admin;
+    }
+    
+    private BasicDBObject insertBinding() {
+        BasicDBObject docInput = new BasicDBObject();
+        docInput.put("destination", "ei-domain.eiffel-intelligence.messageConsumer.durable");
+        docInput.put("destinationType", "QUEUE");
+        docInput.put("exchange", "ei-binding-keys");
+        docInput.put("bindingKeys", BINDING_KEY_1);
+        docInput.put("arg", null);
+        mongoDBHandler.insertDocument(dataBaseName, collectionName, docInput.toString());
+        return docInput;
+    }
+
+    private ArrayList<String> compareAndRemoveBindings() {
+        ArrayList<Binding> listBinding = new ArrayList<Binding>();
+        listBinding.add(new Binding("ei-domain.eiffel-intelligence.messageConsumer.durable", DestinationType.QUEUE,
+                "ei-binding-keys", BINDING_KEY_2, null));
+        List<String> allObjects = mongoDBHandler.getAllDocuments(dataBaseName, collectionName);
+        ArrayList<String> removedBinding = new ArrayList<String>();
+
+        JSONObject dbBinding = new JSONObject(allObjects.get(0));
+        String mongoDbBinding = dbBinding.getString("bindingKeys");
+        if (!(listBinding.contains(mongoDbBinding))) {
+            removedBinding.add(mongoDbBinding);
+            listBinding.remove(0);
+            String condition = "{\"bindingKeys\": /.*" + mongoDbBinding + "/}";
+            mongoDBHandler.dropDocument(dataBaseName, collectionName, condition);
+        }
+        return removedBinding;
     }
 
 }
